Struct three_d_asset::io::RawAssets

pub struct RawAssets(/* private fields */);

Contains raw assets which are usually generated by either the load/load_async functions or the Serialize::serialize function. Can also be constructed manually using the RawAssets::new and RawAssets::insert.

Use the RawAssets::remove or RawAssets::get function to extract the raw byte array for the assets or RawAssets::deserialize to deserialize an asset or RawAssets::save to save the assets.

Implementations

impl RawAssets

pub fn new() -> Self

Constructs a new empty set of raw assets.

pub fn remove(&mut self, path: impl AsRef<Path>) -> Result<Vec<u8>>

Remove and returns the raw byte array for the resource at the given path.

let mut assets = load(&["test_data/test.png"]).unwrap();
let png_bytes = assets.remove("test.png").unwrap();

Note: If no asset has been inserted with the exact same path given as input, then the first asset which path that contains the path given as input is returned. This means the above example can be reduced to:

let png_bytes = load(&["test_data/test.png"]).unwrap().remove("").unwrap();

pub fn get(&self, path: impl AsRef<Path>) -> Result<&[u8]>

Returns a reference to the raw byte array for the resource at the given path.

let mut assets = load(&["test_data/test.png"]).unwrap();
let png_bytes = assets.get("test.png").unwrap();

Note: If no asset has been inserted with the exact same path given as input, then the first asset which path that contains the path given as input is returned. This means the above example can be reduced to:

let png_bytes = load(&["test_data/test.png"]).unwrap().get("").unwrap();

pub fn insert(&mut self, path: impl AsRef<Path>, bytes: Vec<u8>) -> &mut Self

Inserts the given bytes into the set of raw assets. This is useful if you want to add data from an unsuported source and want to use either the RawAssets::deserialize functionality or RawAssets::save functionality.

let mut assets = RawAssets::new();
assets.insert("test.png", png_bytes);
let texture: Texture2D = assets.deserialize("test.png").unwrap();

pub fn extend(&mut self, raw_assets: Self) -> &mut Self

Inserts all of the given raw assets into this set of raw assets.

pub fn deserialize<T: Deserialize>( &mut self, path: impl AsRef<Path> ) -> Result<T>

Deserialize the asset with the given path into a type that implements the Deserialize trait.

let mut assets = load(&["test_data/test.png"]).unwrap();
let texture: Texture2D = assets.deserialize("test.png").unwrap();

Note: If no asset has been inserted with the exact same path given as input, then the first asset which path that contains the path given as input is deserialized. This means the above example can be reduced to:

let texture: Texture2D = load(&["test_data/test.png"]).unwrap().deserialize("").unwrap();

pub fn save(&mut self) -> Result<()>

Saves all of the raw assets to files.