Trait three_d_asset::io::Serialize

pub trait Serialize: Sized {
    // Required method
    fn serialize(&self, path: impl AsRef<Path>) -> Result<RawAssets>;
}

Implemented for assets that can be serialized before being saved (see also save).

Required Methods

fn serialize(&self, path: impl AsRef<Path>) -> Result<RawAssets>

Serialize the asset into a list of raw assets which consist of byte arrays and related path to where they should be saved (see also save). The path given as input is the path to the main raw asset.