Module three_d_asset::material

Contain material asset definitions.

Structs

PbrMaterial

A CPU-side version of a material used for physically based rendering (PBR).

Srgba

Represents a color composed of a red, green and blue component in the sRGB color space. In addition, the alpha value determines the how transparent the color is (0 is fully transparent and 255 is fully opaque).

Texture2D

A CPU-side version of a 2D texture.

Enums

GeometryFunction

The geometry function used in a Cook-Torrance lighting model.

Interpolation

Possible modes of interpolation which determines the texture output between texture pixels.

LightingModel

Lighting models which specify how the lighting is computed when rendering a material. This is a trade-off between how fast the computations are versus how physically correct they look.

NormalDistributionFunction

The normal distribution function used in a Cook-Torrance lighting model.

TextureData

The pixel/texel data for a Texture2D or Texture3D.

Wrapping

Possible wrapping modes for a texture which determines how the texture is applied outside of the [0..1] uv coordinate range.