Module three_d_asset::prelude

Contain the basic types used by the 3D specific data types. Mostly basic math functionality which is an re-export of cgmath.

Structs

AxisAlignedBoundingBox

A bounding box that aligns with the x, y and z axes.

Deg

An angle, in degrees.

Matrix2

A 2 x 2, column major matrix

Matrix3

A 3 x 3, column major matrix

Matrix4

A 4 x 4, column major matrix

Point2

A point in 2-dimensional space.

Point3

A point in 3-dimensional space.

Quaternion

A quaternion in scalar/vector form.

Rad

An angle, in radians.

Srgba

Represents a color composed of a red, green and blue component in the sRGB color space. In addition, the alpha value determines the how transparent the color is (0 is fully transparent and 255 is fully opaque).

Vector2

A 2-dimensional vector.

Vector3

A 3-dimensional vector.

Vector4

A 4-dimensional vector.

f16

A 16-bit floating point type implementing the IEEE 754-2008 standard binary16 a.k.a “half” format.

Traits

Angle

Angles and their associated trigonometric functions.

EuclideanSpace

Points in a Euclidean space with an associated space of displacement vectors.

InnerSpace

Vectors that also have a dot (or inner) product.

Matrix

A column-major matrix of arbitrary dimensions.

MetricSpace

A type with a distance function between values.

One

Defines a multiplicative identity element for Self.

Rotation

A trait for a generic rotation. A rotation is a transformation that creates a circular motion, and preserves at least one point in the space.

Rotation2

A two-dimensional rotation.

Rotation3

A three-dimensional rotation.

SquareMatrix

A column-major major matrix where the rows and column vectors are of the same dimensions.

Transform

A trait representing an affine transformation that can be applied to points or vectors. An affine transformation is one which

Transform2

Transform3

VectorSpace

Vectors that can be added together and multiplied by scalars.

Zero

Defines an additive identity element for Self.

Functions

degrees

Constructs a an angle in degrees.

dot

Dot product of two vectors.

frustum

Create a perspective matrix from a view frustum.

ortho

Create an orthographic projection matrix.

perspective

Create a perspective projection matrix.

radians

Constructs a an angle in radians.

rotation_matrix_from_dir_to_dir

Constructs a rotation matrix that rotates from the source direction to the target direction.

vec2

The short constructor.

vec3

The short constructor.

vec4

The short constructor.

Type Aliases

ColorDeprecated

Represents a color composed of a red, green and blue component in the sRGB color space. In addition, the alpha value determines the how transparent the color is (0 is fully transparent and 255 is fully opaque).

Degrees

A Degrees with f32 data type.

Mat2

A Matrix2 with f32 data type.

Mat3

A Matrix3 with f32 data type.

Mat4

A Matrix4 with f32 data type.

Quat

A Quaternion with f32 data type.

Radians

A Radians with f32 data type.

Vec2

A Vector2 with f32 data type.

Vec3

A Vector3 with f32 data type.

Vec4

A Vector4 with f32 data type.