Struct three_d_asset::prelude::Vector3

#[repr(C)]
pub struct Vector3<S> {
    pub x: S,
    pub y: S,
    pub z: S,
}

A 3-dimensional vector.

This type is marked as #[repr(C)].

Fields
x: S

The x component of the vector.

y: S

The y component of the vector.

z: S

The z component of the vector.

Implementations

impl<S> Vector3<S>

pub const fn new(x: S, y: S, z: S) -> Vector3<S>

Construct a new vector, using the provided values.

pub fn map<U, F>(self, f: F) -> Vector3<U>where F: FnMut(S) -> U,

Perform the given operation on each field in the vector, returning a new point constructed from the operations.

pub fn zip<S2, S3, F>(self, v2: Vector3<S2>, f: F) -> Vector3<S3>where F: FnMut(S, S2) -> S3,

Construct a new vector where each component is the result of applying the given operation to each pair of components of the given vectors.

impl<S> Vector3<S>where S: NumCast + Copy,

pub fn cast<T>(&self) -> Option<Vector3<T>>where T: NumCast,

Component-wise casting to another type.

impl<S> Vector3<S>where S: BaseNum,

pub fn unit_x() -> Vector3<S>

A unit vector in the x direction.

pub fn unit_y() -> Vector3<S>

A unit vector in the y direction.

pub fn unit_z() -> Vector3<S>

A unit vector in the z direction.

pub fn cross(self, other: Vector3<S>) -> Vector3<S>

Returns the cross product of the vector and other.

pub fn extend(self, w: S) -> Vector4<S>

Create a Vector4, using the x, y and z values from this vector, and the provided w.

pub fn truncate(self) -> Vector2<S>

Create a Vector2, dropping the z value.