Type Alias three_d_asset::prelude::Color

pub type Color = Srgba;

๐Deprecated: Renamed to Srgba

Represents a color composed of a red, green and blue component in the sRGB color space. In addition, the alpha value determines the how transparent the color is (0 is fully transparent and 255 is fully opaque).

Aliased Type

struct Color {
    pub r: u8,
    pub g: u8,
    pub b: u8,
    pub a: u8,
}

Fields
r: u8

Red component

g: u8

Green component

b: u8

Blue component

a: u8

Alpha component

Implementations

impl Srgba

pub const fn new(r: u8, g: u8, b: u8, a: u8) -> Self

Creates a new sRGBA color with the given values.

pub const fn new_opaque(r: u8, g: u8, b: u8) -> Self

Creates a new sRGB color with the given red, green and blue values and an alpha value of 255.

pub fn to_linear_srgb(&self) -> Vec4

Returns the color in linear sRGB color space.

pub const RED: Self = _

Opaque red

pub const GREEN: Self = _

Opaque green

pub const BLUE: Self = _

Opaque blue

pub const WHITE: Self = _

Opaque white

pub const BLACK: Self = _

Opaque black