Type Alias three_d_asset::prelude::Mat4

pub type Mat4 = Matrix4<f32>;

A Matrix4 with f32 data type.

Aliased Type

struct Mat4 {
    pub x: Vector4<f32>,
    pub y: Vector4<f32>,
    pub z: Vector4<f32>,
    pub w: Vector4<f32>,
}

Fields
x: Vector4<f32>

The first column of the matrix.

y: Vector4<f32>

The second column of the matrix.

z: Vector4<f32>

The third column of the matrix.

w: Vector4<f32>

The fourth column of the matrix.

Implementations

impl<S> Matrix4<S>

pub const fn new( c0r0: S, c0r1: S, c0r2: S, c0r3: S, c1r0: S, c1r1: S, c1r2: S, c1r3: S, c2r0: S, c2r1: S, c2r2: S, c2r3: S, c3r0: S, c3r1: S, c3r2: S, c3r3: S ) -> Matrix4<S>

Create a new matrix, providing values for each index.

pub const fn from_cols( c0: Vector4<S>, c1: Vector4<S>, c2: Vector4<S>, c3: Vector4<S> ) -> Matrix4<S>

Create a new matrix, providing columns.

impl<S> Matrix4<S>where S: BaseFloat,

pub fn from_translation(v: Vector3<S>) -> Matrix4<S>

Create a homogeneous transformation matrix from a translation vector.

pub fn from_scale(value: S) -> Matrix4<S>

Create a homogeneous transformation matrix from a scale value.

pub fn from_nonuniform_scale(x: S, y: S, z: S) -> Matrix4<S>

Create a homogeneous transformation matrix from a set of scale values.

pub fn look_at_dir( eye: Point3<S>, dir: Vector3<S>, up: Vector3<S> ) -> Matrix4<S>

๐Deprecated: Use Matrix4::look_to_rh

Create a homogeneous transformation matrix that will cause a vector to point at dir, using up for orientation.

pub fn look_to_rh(eye: Point3<S>, dir: Vector3<S>, up: Vector3<S>) -> Matrix4<S>

Create a homogeneous transformation matrix that will cause a vector to point at dir, using up for orientation.

pub fn look_to_lh(eye: Point3<S>, dir: Vector3<S>, up: Vector3<S>) -> Matrix4<S>

Create a homogeneous transformation matrix that will cause a vector to point at dir, using up for orientation.

pub fn look_at(eye: Point3<S>, center: Point3<S>, up: Vector3<S>) -> Matrix4<S>

๐Deprecated: Use Matrix4::look_at_rh

Create a homogeneous transformation matrix that will cause a vector to point at center, using up for orientation.

pub fn look_at_rh( eye: Point3<S>, center: Point3<S>, up: Vector3<S> ) -> Matrix4<S>

Create a homogeneous transformation matrix that will cause a vector to point at center, using up for orientation.

pub fn look_at_lh( eye: Point3<S>, center: Point3<S>, up: Vector3<S> ) -> Matrix4<S>

Create a homogeneous transformation matrix that will cause a vector to point at center, using up for orientation.

pub fn from_angle_x<A>(theta: A) -> Matrix4<S>where A: Into<Rad<S>>,

Create a homogeneous transformation matrix from a rotation around the x axis (pitch).

pub fn from_angle_y<A>(theta: A) -> Matrix4<S>where A: Into<Rad<S>>,

Create a homogeneous transformation matrix from a rotation around the y axis (yaw).

pub fn from_angle_z<A>(theta: A) -> Matrix4<S>where A: Into<Rad<S>>,

Create a homogeneous transformation matrix from a rotation around the z axis (roll).

pub fn from_axis_angle<A>(axis: Vector3<S>, angle: A) -> Matrix4<S>where A: Into<Rad<S>>,

Create a homogeneous transformation matrix from an angle around an arbitrary axis.

The specified axis must be normalized, or it represents an invalid rotation.

pub fn is_finite(&self) -> bool

Are all entries in the matrix finite.

impl<S> Matrix4<S>where S: NumCast + Copy,

pub fn cast<T>(&self) -> Option<Matrix4<T>>where T: NumCast,

Component-wise casting to another type