Type Alias three_d_asset::prelude::Vec4

pub type Vec4 = Vector4<f32>;

A Vector4 with f32 data type.

Aliased Type

struct Vec4 {
    pub x: f32,
    pub y: f32,
    pub z: f32,
    pub w: f32,
}

Fields
x: f32

The x component of the vector.

y: f32

The y component of the vector.

z: f32

The z component of the vector.

w: f32

The w component of the vector.

Implementations

impl<S> Vector4<S>

pub const fn new(x: S, y: S, z: S, w: S) -> Vector4<S>

Construct a new vector, using the provided values.

pub fn map<U, F>(self, f: F) -> Vector4<U>where F: FnMut(S) -> U,

Perform the given operation on each field in the vector, returning a new point constructed from the operations.

pub fn zip<S2, S3, F>(self, v2: Vector4<S2>, f: F) -> Vector4<S3>where F: FnMut(S, S2) -> S3,

Construct a new vector where each component is the result of applying the given operation to each pair of components of the given vectors.

impl<S> Vector4<S>where S: NumCast + Copy,

pub fn cast<T>(&self) -> Option<Vector4<T>>where T: NumCast,

Component-wise casting to another type.

impl<S> Vector4<S>where S: BaseNum,

pub fn unit_x() -> Vector4<S>

A unit vector in the x direction.

pub fn unit_y() -> Vector4<S>

A unit vector in the y direction.

pub fn unit_z() -> Vector4<S>

A unit vector in the z direction.

pub fn unit_w() -> Vector4<S>

A unit vector in the w direction.

pub fn truncate(self) -> Vector3<S>

Create a Vector3, dropping the w value.

pub fn truncate_n(&self, n: isize) -> Vector3<S>

Create a Vector3, dropping the nth element.