Module three_d_asset::volume

Contain volume asset definitions.

Re-exports

pub use crate::prelude::*;

Structs

Texture3D

A CPU-side version of a 3D texture.

VoxelGrid

Volume data consisting of voxel data inside a cube.

Enums

Interpolation

Possible modes of interpolation which determines the texture output between texture pixels.

TextureData

The pixel/texel data for a Texture2D or Texture3D.

Wrapping

Possible wrapping modes for a texture which determines how the texture is applied outside of the [0..1] uv coordinate range.